refactor(cli): colocate MCP context fixture
