+++ packages/cli/AGENTS.md
@@ -90,7 +90,7 @@ The MCP server (`mcp.ts`) exposes a curated subset of commands as MCP tools via 
 - `process.stderr.write(...)` → `ctx.process.stderr.write(...)`
 - `process.exit(code)` → `ctx.process.exit(code)`
 
-Use the shared `ActionCtx` type from `src/lib/action-ctx.ts` — a structural subtype of `GokeExecutionContext` that captures exactly the surface actions consume. `createCapturedCtx()`, `expectCapturedToInclude()`, and `MockProcessExit` remain in `src/__test__/mcp-ctx-fixture.ts` for test use only. Action bodies are exported as named functions (e.g., `gatewayStatusAction`, `cliproxyKeysListAction`) and the `.action(...)` callback delegates to them, so Tier-2 tests invoke actions directly with `createCapturedCtx()`.
+Use the shared `ActionCtx` type from `src/lib/action-ctx.ts` — a structural subtype of `GokeExecutionContext` that captures exactly the surface actions consume. `createCapturedCtx()`, `expectCapturedToInclude()`, and `MockProcessExit` remain in `src/lib/mcp-ctx-fixture.ts` for test use only. Action bodies are exported as named functions (e.g., `gatewayStatusAction`, `cliproxyKeysListAction`) and the `.action(...)` callback delegates to them, so Tier-2 tests invoke actions directly with `createCapturedCtx()`.
 
 **Current MCP surface = 9 read-only tools (all Mode A): the app status tools, the unified `status` tool, plus `cliproxy models`.** The `agent setup`, `agent storage`, and `agent storage teardown` commands are **source-gated out of `MCP_ALLOWLIST`** because they mutate GitHub/AWS configuration and require operator-controlled files, credentials, or environment approval. They are available through the direct CLI only. The existing mutating/secret-disclosing commands remain source-gated as well. The Mode B/C and mutation-verification guidance below documents the capture _contract_ and remains the bar for any future capturable command.
