+++ packages/cli/src/commands/broker/status.test.ts
@@ -1,6 +1,6 @@
 import {afterEach, beforeEach, describe, expect, it} from 'bun:test'
 
-import {createCapturedCtx, expectCapturedToInclude, MockProcessExit} from '../../__test__/mcp-ctx-fixture'
+import {createCapturedCtx, expectCapturedToInclude, MockProcessExit} from '../../lib/mcp-ctx-fixture'
 import {brokerStatusAction, checkBrokerHealth, getBrokerStatusSummary} from './status'
 
 const originalFetch = globalThis.fetch
