+++ packages/cli/src/commands/cliproxy/keys.test.ts
@@ -1,6 +1,6 @@
 import {afterEach, beforeEach, describe, expect, it} from 'bun:test'
 
-import {createCapturedCtx, expectCapturedToInclude} from '../../__test__/mcp-ctx-fixture'
+import {createCapturedCtx, expectCapturedToInclude} from '../../lib/mcp-ctx-fixture'
 import {cliproxyKeysAddAction, cliproxyKeysListAction, cliproxyKeysRemoveAction, toStringArray} from './keys'
 
 const originalFetch = globalThis.fetch
