+++ packages/cli/src/commands/cliproxy/models.test.ts
@@ -1,7 +1,7 @@
 import {afterEach, beforeEach, describe, expect, it} from 'bun:test'
 import {goke} from 'goke'
 
-import {createCapturedCtx, MockProcessExit} from '../../__test__/mcp-ctx-fixture'
+import {createCapturedCtx, MockProcessExit} from '../../lib/mcp-ctx-fixture'
 import {cliproxyModelsAction, registerCliproxyModels} from './models'
 
 const originalFetch = globalThis.fetch
