+++ packages/cli/src/commands/keeweb/status.test.ts
@@ -1,6 +1,6 @@
 import {afterEach, beforeEach, describe, expect, it, mock, spyOn} from 'bun:test'
 
-import {createCapturedCtx, expectCapturedToInclude} from '../../__test__/mcp-ctx-fixture'
+import {createCapturedCtx, expectCapturedToInclude} from '../../lib/mcp-ctx-fixture'
 import {
   checkContentHash,
   checkHttpReachability,
